# CostModel.ComputeAllocation using query instead of query_range