--- a/pkg/costmodel/allocation.go
+++ b/pkg/costmodel/allocation.go
@@ -47,6 +47,8 @@ const (
 	queryFmtJobLabels             = `sum(avg_over_time(kube_pod_owner{owner_kind="Job"}[%s]%s)) by (pod, owner_name, namespace ,cluster_id)`
 )
 
+// TODO niko/computeallocation idle minutes = 1?
+
 // ComputeAllocation uses the CostModel instance to compute an AllocationSet
 // for the window defined by the given start and end times. The Allocations
 // returned are unaggregated (i.e. down to the container level).