Required for https://github.com/kubecost/kubecost-cost-model/pull/332

## Changes
- Implemented new `ComputeAllocation` function on `CostModel` so that it can directly serve as an ETL source
- Compute CPU, RAM, and Total efficiency on the fly using the now-passed-through Request and Usage averages for each resource

## Testing
In general, numbers are looking very accurate using [Reports drill-down](https://github.com/AjayTripathy/cost-analyzer-frontend/pull/160) to test. They look very similar in most cases to the existing data, too, but in other cases they can be off enough that it has been somewhat laborious to try to see which is more accurate.

### Accuracy
Tested manually comparing results from this image at `kubecost.nikovacevic.io` to results from `staging.kubecost.nikovacevic.io`. The following numbers come from "yesterday" as of this writing, i.e. 2021-02-18 to 2021-02-19

#### Cluster: cluster-one
| | old | new |
|-|-|-|
| minutes | 1440 | 1440 |
| cpu core hrs | 58.063422 | 55.74458 |
| cpu cores | 2.41930 | 2.32269 |
| cpu cost | 1.01553 | 1.22162 |
| cpu efficiency | 0.08384 | 0.04415 |
| ram gib hrs | 77.44820 | 75.11350 |
| ram gib | 3.22700 | 3.12972 |
| ram cost | 0.18471 | 0.22078 |
| ram efficiency | 1.01892 | 0.43509 |
| pv gib hrs | 3475 | 3167.63072 |
| pv gib | 144.79166 | 131.98461 |
| pv cost | 0.19108 | 0.17356 |
| total cost | 1.39133 | 1.61597 |
| total efficiency | 0.33306 | 0.09510 |

#### Namespace: kubecost
| | old | new |
|-|-|-|
| minutes | 1440 | 1440 |
| cpu core hrs | 8.27975 | 7.88554 |
| cpu cores | 0.34498 | 0.32856 |
| cpu cost | 0.13576 | 0.17243 |
| cpu efficiency | 0.13677 | 0.00977 |
| ram gib hrs | 11.99318 | 11.43355 |
| ram gib | 0.49971 | 0.47639 |
| ram cost | 0.02542 | 0.03347 |
| ram efficiency | 1.41941 | 0.21537 |
| pv gib hrs | 1650 | 1583.46666 |
| pv gib | 68.75 | 65.97777 |
| pv cost | 0.09075 | 0.08676 |
| total cost | 0.25193 | 0.29267 |
| total efficiency | 0.59183 | 0.03633 |

#### Controller: deployment/kubecost-cost-analyzer
| | old | new |
|-|-|-|
| minutes | 1440 | 1440 |
| cpu core hrs | 7.73450 | 7.43483 |
| cpu cores | 0.32227 | 0.30979 |
| cpu cost | 0.12735 | 0.16260 |
| cpu efficiency | 0.01424 | 0.01037 |
| ram gib hrs | 4.24561 | 4.07669 |
| ram gib | 0.17690 | 0.16986 |
| ram cost | 0.00935 | 0.01195 |
| ram efficiency | 0.42731 | 0.60208 |
| pv gib hrs | 798.93333 | 767.46667 |
| pv gib | 33.28889 | 31.97778 |
| pv cost | 0.04394 | 0.04205 |
| total cost | 0.18065 | 0.21660 |
| total efficiency | 0.03744 | 0.04903 |

### Performance
Currently working on it on the thanos-sim cluster...

Early results from raw queries suggested the following before/after for the allocation queries:
```
RAM allocation (24h)
- count: 114 -> 114
- time: 1000ms -> 120ms
- size: 3.6MB -> 23kB

CPU allocation (24h)
- count: 112 -> 112
- time: 1200ms -> 160ms
- size: 4.1MB -> 23kB
```