@nikovacevic 2021-03-02
Closing in favor of https://github.com/kubecost/cost-model/pull/719